requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
celery==3.1.18
Django==1.10.6
django-celery==3.2.1
httplib2==0.10.3
kombu==3.0.37
oauth2==1.9.0.post1
packaging==16.8
pyenchant==1.6.8
pyparsing==2.2.0
pytz==2017.2
redis==2.10.3
six==1.10.0
vine==1.1.3
